Make specs/rubo pass
Having brought this in from a Rails app, make specs work by:

- Using ActiveRecord locally with an in-memory SQLite database (make
  sure to clear the DB between examples)
- Appending to I18n.load_paths such that we have our default SVG prev
  and next buttons in English
- Excluding all the nasty formatting (this is based on
  pagy/extras/bootstrap, so ape that formatting for better or worse)

The relevant commit messages in the original repo are as follows:

https://design-system.service.gov.uk/components/pagination/ is the
reference.

Note that this has issues compared to that reference due to a clash
in philosophy. Pagy is micro-optimised and no longer has templates,
rather just modules with strings.

As a result, pagy_anchor needs attention in order to be able to get the
SVG arrows in, as it's a rather simple lambda. Perhaps for prev and
next links, we use a small alternative.

In addition there is an issue with the current page. It is black on
dark blue. I cannot currently work out why the selector

```
nav > ul > li.govuk-pagination__item.govuk-pagination__item--current > a
```

is not being matched.

Delete the manual pagy styling.

Pagy uses its own I18n for micro-optimisation reasons. It's militantly
unaware of our particular multitenant setup. Even the Pagy I18n extra
is also unaware of this.

Replace `pagy_t` calls in our prev/next links with ::I18n.t calls, and
add a full GOV.UK set of HTML to en.yml to provide these snippets.